correlation: move around some modules
