--- correlation/src/dispatcher/handlers/mod.rs
@@ -1,10 +1,0 @@
-// Copyright (c) 2016 Tibor Benke <ihrwein@gmail.com>
-//
-// Licensed under the Apache License, Version 2.0
-// <LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0> or the MIT
-// license <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
-// option. All files in the project carrying such notice may not be copied,
-// modified, or distributed except according to those terms.
-
-pub mod message;
-pub mod timer;
